rename module posterior -> posteriors
