+++ doc/source/notebooks/advanced/variational_fourier_features.pct.py
@@ -236,10 +236,10 @@ def gauss_kl_vff(q_mu, q_sqrt, K):
 
 
 # %%
-import gpflow.posterior
+import gpflow.posteriors
 
 
-class VFFPosterior(gpflow.posterior.AbstractPosterior):
+class VFFPosterior(gpflow.posteriors.AbstractPosterior):
     def _conditional_fused(self, Xnew, full_cov, full_output_cov):
         """
         Xnew is a tensor with the points of the data or minibatch, shape N x D
@@ -359,7 +359,7 @@ class VFFPosterior(gpflow.posterior.AbstractPosterior):
 # We now have to register our Posterior object:
 
 # %%
-@gpflow.posterior.get_posterior_class.register(gpflow.kernels.Kernel, FourierFeatures1D)
+@gpflow.posteriors.get_posterior_class.register(gpflow.kernels.Kernel, FourierFeatures1D)
 def _get_posterior_vff(kernel, inducing_variable):
     return VFFPosterior
